feat(ui): add border width and radius controls to flag theme extension
- Introduced BorderFlagThemeControllerExtension for managing border width
  and radius.
- Updated FlagThemeController to expose default border radius.
- Modified relevant UI components to utilize new border controls.
