+++ packages/world_countries/example/lib/theme/flag_theme_controller.dart
@@ -12,12 +12,12 @@ class FlagThemeController extends ChangeNotifier {
   );
 
   // ignore: avoid-explicit-type-declaration, vs specify_nonobvious_property_types.
-  static const BorderRadius _defaultBorderRadius = BorderRadius.zero;
+  static const BorderRadius defaultBorderRadius = BorderRadius.zero;
   static const _defaultTheme = FlagThemeData(
     decorationPosition: DecorationPosition.background,
     aspectRatio: FlagConstants.defaultAspectRatio,
     decoration: BoxDecoration(
-      borderRadius: _defaultBorderRadius,
+      borderRadius: defaultBorderRadius,
       border: Border.fromBorderSide(defaultBorder),
     ),
   );
@@ -26,9 +26,6 @@ class FlagThemeController extends ChangeNotifier {
 
   double get aspectRatio =>
       _theme.specifiedAspectRatio ?? FlagConstants.defaultAspectRatio;
-  double get borderRadius =>
-      _theme.decoration?.borderRadius?.resolve(null).topRight.x ?? 0;
-  double? get borderWidth => _theme.decoration?.border?.top.width;
   FlagThemeData get theme => _theme;
 
   void reset() => _setTheme();
@@ -36,33 +33,6 @@ class FlagThemeController extends ChangeNotifier {
   set aspectRatio(double? value) => _setTheme(
     _theme.copyWith(aspectRatio: value ?? _defaultTheme.aspectRatio),
   );
-
-  set borderRadius(double? value) => setDecoration(
-    _theme.decoration?.copyWith(
-      borderRadius: value == null
-          ? _defaultBorderRadius
-          : BorderRadius.all(Radius.circular(value)),
-    ),
-  );
-
-  set borderWidth(double? value) {
-    final width = value ?? 1;
-    final decoration = _theme.decoration;
-    if (width > 0) {
-      setDecoration(
-        decoration?.copyWith(
-          border: Border.fromBorderSide(defaultBorder.copyWith(width: width)),
-        ),
-      );
-    } else {
-      setDecoration(
-        BoxDecoration(
-          borderRadius: decoration?.borderRadius,
-          boxShadow: decoration?.boxShadow,
-        ),
-      );
-    }
-  }
 
   void setDecoration(BoxDecoration? decoration) =>
       _setTheme(_theme.copyWith(decoration: decoration));
