+++ packages/world_countries/example/lib/ui/settings/widgets/appbar/scaled_country_flag.dart
@@ -1,6 +1,7 @@
 import "package:flutter/material.dart";
 import "package:world_countries/world_countries.dart";
 
+import "../../../../helpers/extensions/border_flag_theme_controller_extension.dart";
 import "../../../../theme/flag_theme_controller.dart";
 import "../appbar/sliver_settings_app_bar.dart";
 
@@ -24,32 +25,33 @@ class ScaledCountryFlag extends StatelessWidget {
       final decoration = _controller.theme.decoration;
       final shadow = decoration?.boxShadow?.firstOrNull;
       final borderWidth = _controller.borderWidth;
-      final scaledDecoration = decoration?.copyWith(
-        borderRadius: BorderRadius.all(
-          Radius.circular(_controller.borderRadius * scale),
-        ),
-        border: borderWidth == null
-            ? null
-            : Border.all(
-                strokeAlign: FlagThemeController.defaultBorder.strokeAlign,
-                width: borderWidth * scale,
-                color:
-                    decoration.border?.top.color ?? // TODO!
-                    FlagThemeController.defaultBorder.color,
-              ),
-        boxShadow: [
-          ?shadow?.copyWith(
-            blurRadius: shadow.blurRadius * scale,
-            spreadRadius: shadow.spreadRadius * scale,
-            offset: Offset(shadow.offset.dx * scale, shadow.offset.dy * scale),
-          ),
-        ],
-      );
+      final color = decoration?.color;
 
       return SliverSettingsAppBar(
         (isSimplified ? fallback : _flag).copyWith(
           aspectRatio: _controller.theme.specifiedAspectRatio,
-          decoration: scaledDecoration,
+          decoration: decoration?.copyWith(
+            borderRadius: BorderRadius.all(
+              Radius.circular(_controller.borderRadius * scale),
+            ),
+            border: borderWidth == null
+                ? null
+                : Border.all(
+                    color: color ?? FlagThemeController.defaultBorder.color,
+                    strokeAlign: FlagThemeController.defaultBorder.strokeAlign,
+                    width: borderWidth * scale,
+                  ),
+            boxShadow: [
+              ?shadow?.copyWith(
+                blurRadius: shadow.blurRadius * scale,
+                spreadRadius: shadow.spreadRadius * scale,
+                offset: Offset(
+                  shadow.offset.dx * scale,
+                  shadow.offset.dy * scale,
+                ),
+              ),
+            ],
+          ),
         ),
         onReset: _controller.reset,
       );
