+++ packages/world_countries/example/lib/ui/settings/widgets/sliders/border_width_slider.dart
@@ -1,5 +1,6 @@
 import "package:flutter/material.dart";
 
+import "../../../../helpers/extensions/border_flag_theme_controller_extension.dart";
 import "settings_tile_slider.dart";
 
 /// Slider controlling the border stroke width.
